+++ lib/graphql/stitching/supergraph/source_directive.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module GraphQL::Stitching
+  class Supergraph
+    class SourceDirective < GraphQL::Schema::Directive
+      graphql_name "source"
+      locations FIELD_DEFINITION
+      argument :location, String, required: true
+      repeatable true
+    end
+  end
+end
